backend/requirements.txt
django>=4.2,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-filter>=23.0
django-cors-headers>=4.3
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
whitenoise>=6.6
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn>=1.4.0
numpy>=1.26.0
Pillow>=10.0
